+++ docs/authorization.rst
@@ -1,0 +1,12 @@
+Authorization
+=============
+
+.. automodule:: virtool.authorization.__init__
+    :members:
+
+Client
+------
+
+.. automodule:: virtool.authorization.client
+    :members:
+
