+++ docs/data.rst
@@ -1,0 +1,27 @@
+Data Layer
+==========
+
+.. automodule:: virtool.data.__init__
+    :members:
+
+Transforms
+----------
+
+.. automodule:: virtool.data.transforms
+
+
+API
++++
+
+.. autoclass:: virtool.data.transforms.AbstractTransform
+    :members:
+
+.. autofunction:: virtool.data.transforms.apply_transforms
+
+Errors
+------
+
+.. automodule:: virtool.data.errors
+    :members:
+
+
