+++ docs/index.rst
@@ -1,0 +1,14 @@
+Welcome to Virtool's documentation!
+===================================
+
+This is great!
+
+.. toctree::
+   :maxdepth: 2
+
+   authorization
+   data
+   types
+
+   testing/faking
+   testing/fixtures
