+++ docs/testing/faking.rst
@@ -1,0 +1,5 @@
+Faking
+======
+
+In Virtool, faking refers to creating fake resources for testing and development.
+
