+++ docs/testing/fixtures.rst
@@ -1,0 +1,36 @@
+Fixtures
+========
+
+Authorization
+-------------
+
+.. automodule:: tests.fixtures.authorization
+    :members:
+
+
+Client
+------
+
+.. automodule:: tests.fixtures.client
+    :members:
+
+Data
+----
+
+.. automodule:: tests.fixtures.data
+    :members:
+
+
+Migration
+---------
+
+.. automodule:: tests.fixtures.migration
+    :members:
+
+Postgres
+--------
+
+.. automodule:: tests.fixtures.pg
+    :members:
+
+
