+++ docs/transforms.rst
@@ -1,0 +1,13 @@
+Transforms
+==========
+
+.. automodule:: virtool.data.transforms
+
+
+API
+---
+
+.. autoclass:: virtool.data.transforms.AbstractTransform
+    :members:
+
+.. autofunction:: virtool.data.transforms.apply_transforms
