+++ docs/types.rst
@@ -1,0 +1,5 @@
+Types
+=====
+
+.. automodule:: virtool.types
+    :members:
