+++ pyproject.toml
@@ -77,6 +77,9 @@ pre-commit = "^3.3.3"
 black = "^23.3.0"
 pytest-rerunfailures = "^10.2"
 pytest-asyncio = "^0.21.0"
+sphinx = "^7.0.1"
+sphinx-autobuild = "^2021.3.14"
+sphinx-autofixture = "^0.4.0"
 
 [tool.pytest.ini_options]
 addopts = "--capture tee-sys"
