+++ tests/fixtures/__init__.py
@@ -1,0 +1,3 @@
+"""
+Pytest fixtures for Virtool.
+"""
