+++ tests/fixtures/data.py
@@ -1,8 +1,9 @@
+"""Fixtures for working with the Virtool data layer."""
 import pytest
 from aiohttp import ClientSession
 from aioredis import Redis
 from sqlalchemy.ext.asyncio import AsyncEngine
-from virtool.data.factory import create_data_layer
+from virtool.data.layer import create_data_layer
 from virtool.tasks.client import TasksClient
 from virtool.tasks.data import TasksData
 
@@ -11,6 +12,18 @@ from virtool.tasks.data import TasksData
 def data_layer(
     authorization_client, mongo, config, mocker, pg: AsyncEngine, redis: Redis
 ):
+    """
+    A complete data layer backed by testing instances of MongoDB, PostgreSQL, OpenFGA,
+    and Redis.
+
+    Example:
+
+    .. code-block:: python
+
+        async def test_example(data_layer: DataLayer):
+            await data_layer.samples.create(...)
+
+    """
     return create_data_layer(
         authorization_client, mongo, pg, config, mocker.Mock(spec=ClientSession), redis
     )
