+++ virtool/administrators/api.py
@@ -17,8 +17,10 @@ from virtool.administrators.oas import (
     RunActionRequest,
 )
 from virtool.api.response import NotFound, json_response
-from virtool.authorization.client import AuthorizationClient
-from virtool.authorization.utils import get_authorization_client_from_req
+from virtool.authorization.client import (
+    AuthorizationClient,
+    get_authorization_client_from_req,
+)
 from virtool.data.errors import ResourceNotFoundError, ResourceError
 from virtool.data.utils import get_data_from_req
 from virtool.http.policy import policy, AdministratorRoutePolicy
