+++ virtool/authorization/__init__.py
@@ -1,0 +1,10 @@
+"""
+
+Authorization is the process of determining whether a user is allowed to perform an
+action.
+
+The Virtool :class:`.AuthorizationClient` is an abstraction over OpenFGA that provides
+a simple interface for checking whether a user has a given permission on a given
+resource.
+
+"""
